+++ app/Filament/Resources/MobileCrashReports/Pages/ViewMobileCrashReport.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Resources\MobileCrashReports\Pages;
+
+use App\Filament\Resources\MobileCrashReports\MobileCrashReportResource;
+use Filament\Resources\Pages\ViewRecord;
+
+class ViewMobileCrashReport extends ViewRecord
+{
+    protected static string $resource = MobileCrashReportResource::class;
+}